Another fix for revision info.
The nightly tarball script was correctly writing version.py,
but the install script was overwriting it.  No longer.

--HG--
extra : convert_revision : svn%3A3ed01bd8-26fb-0310-9e4c-ca1a4053419f/networkx/trunk%401343

+++ networkx/release.py
@@ -16,6 +16,11 @@ def write_versionfile():
     """Creates a file containing version information."""
     base = os.path.split(__file__)[0]
     versionfile = os.path.join(base, 'version.py')
+    if revision is None and os.path.isfile(versionfile):
+        # Unable to get revision info, so probably not in an SVN directory
+        # If a version.py already exists, let's not overwrite it.
+        # Useful mostly for nightly tarballs.
+        return
     fh = open(versionfile, 'w')
     text = '''"""
 Version information for NetworkX, created during installation.
